New stat_utils file for statistical test utilities
